+++ CHANGELOG.rst
@@ -9,6 +9,12 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 Unreleased
 **********
 
+Added
+=====
+
+* Accepts newline separated data and puts each record to the stream
+  individually.
+
 
 *******************
 v0.1.0 - 2015-10-12